src/requirements.txt
line-bot-sdk
boto3
